refactor: finish video filenames and conditional cleanup
